fixed parallel test
